+++ react/__mocks__/vtex.styleguide.js
@@ -1,7 +1,6 @@
 import React from 'react'
-import Button from '@vtex/styleguide/lib/Button'
-import Input from '@vtex/styleguide/lib/Input'
 import Tag from '@vtex/styleguide/lib/Tag'
 import ButtonPlain from '@vtex/styleguide/lib/ButtonPlain'
+import InputButton from '@vtex/styleguide/lib/InputButton'
 
-module.exports = { Button, ButtonPlain, Input, Tag }
+module.exports = { ButtonPlain, InputButton, Tag }